requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.7
prometheus-client==0.20.0
pydantic==2.8.2
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Performance testing dependencies
aiohttp>=3.8.0
matplotlib>=3.5.0
numpy>=1.21.0

# OpenTelemetry dependencies for distributed tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-requests==0.42b0